Update patch set 5
Patch Set 5:

I think this is a wrong integration pattern. The generated MXBean is sufficient interface specification and the datastore implementations should expose a package-private implementation of the bean, which will work closely with the innards of the implementation.

This will prevent the need to have a dedicated, uber-thread-safe, generic DTO and instead will allow the implementation to gather and expose the counters in a manner which affects the performance characteristics of its normal operation the least.

Patch-set: 5
Label: Verified=0