Update patch set 6
Patch Set 6:

My intent was to separate the MBean stuff from the impl (i.e. InMemoryDOMDataStore) and let the component using the IMDS create, organize and register MBeans as they see fit. This is not wrong - it's one way to do it. For the CDS, the IMDS is part of a shard and the IMDS stats are incorporated with the shard stats.  So having the IMDS create an MXBean internally and register/unregister it, as it seems like you're saying, wouldn't play well with the CDS.

At this point, what I submitted works and there's little time left for Helium so, unless there's something functionally wrong, I'd rather focus on getting the stats in. That's the important thing right now, whether or not the "integration" pattern is right or wrong is a matter of opinion and I don't want to get bogged down by theoretical stuff. If you feel it isn't right then please create a bug for post-Helium and propose another solution. Or do it for Helium if you have the time.

Patch-set: 6
Label: Verified=0