Update patch set 8
Patch Set 8:

Unfortunately we cannot, as it has a direct dependency on sun.misc, which is not exposed by OSGi containers by default -- meaning that that is a post-Helium item. As such it has been removed from the util package.

This is one of the most performance-critical areas we have, so careful design and weighing of options is necessary.

As for the general approach, my immediate (without analyzing the specific statistics or where to record them) would be have each individual statistic recorded at the site where it poses the minimal run-time overhead, with a reasonable accuracy/performance trade-off. Then create a class implementing the MXBean interface, which will know where to send those queries. Bias the collection mechanism heavily against reading the stats -- sacrificing accuracy, 'single view' of multiple stats members and access speed as needed. At the end of the day statistics are exactly what Mark Twain describe them to be.

I understand your concern about being bogged down on theory. Can you run this patchset against the datastore performance benchmark (opendaylight/md-sal/benchmark-data-store), so we get a less theoretical view of what the cost is going to be?

Patch-set: 8
Label: Verified=0