Bug 1879: Speed-up feature install.
Opendaylight-Karaf distribution used default configuration for PAX,
which is unsuitable for projects with lot of snapshots and
actual copy of repository present in system directory.

Karaf used default configuration for resolving bundles which is
following:
  1. maven central
  2. springsource
  3. apache
  4. local folders (/system)

Which means each and every opendaylight component was looked
up in these repositories before looking in local folder.
This patch proposes change to this order:

  1. local folders
  2. maven central
  3. springsource
  4. apache

Change-Id: I7adb0195153271bbb3505b3e2e6202e105f27b7c
Signed-off-by: Tony Tkacik <ttkacik@cisco.com>