Update patch set 1
Patch Set 1:

Are you suggesting that we should append ApplyState to the persistent journal for each ReplicatedLogEntry that has been applied to the state?
If yes, wont it increase the persistent-journal size , worst case by twice?

And i didnt get how the check for replicatedLog.isPresent() can be avoided even if we use ApplyState. To avoid this check we can omit storing the unapplied entries as part of the snapshot, and instead rely on the replaying of journal to populate the in-mem log.

Patch-set: 1
Reviewer: Gerrit User 1676 <1676@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0