Bug 1948: Separate out restconf features.
In order to avoid a maven project cycle in solving
Bug 1948, we need to separate restconf features.

Note, this is a first step, suffixing everything
with -new.  Subsquently, after everywhere using
odl-restconf has been fixed to use this new repo,
we will deprecate the ones in the mdsal features.xml
and rename these to not have the -new.

Change-Id: I80451e055c071b30a1d0a83ec5a5c14c41afb586
Signed-off-by: Ed Warnicke <eaw@cisco.com>