Update patch set 3
Patch Set 3: Code-Review-1

I'm confused :-) Is this a new file as opposed to adding a new connector to an existing web.xml file?

Shouldn't the trusted proxy connector be included (it not the same as the admin port)?. Perhaps the template for it should be included inside a comment (so it's not activated by default). That way an admin can uncomment the boilerplate template and tweak a value if needed. I think it's good to help admins by not making them search out the solution, let them find the info in the file with instructions on what to do.

Patch-set: 3
Reviewer: Gerrit User 2084 <2084@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=-1