Update patch set 1
Patch Set 1:

Also, I realize we are short on time, so this is not a MUST, but it would be a little easier to understand what is going on if you take the inner classes out of the main class. That way the synchronization and member access will be a little easier to follow.

Again, not a blocking issue, but may help to isolate the cause of the timing errors.

Patch-set: 1