Update patch set 6
Patch Set 6:

I didn't consider making it an actor. I view it as a helper class for Shard and it's only used during recovery so I'm not sure we need the overhead of defining a new actor and messages. If we did make it an actor we'd have to refactor it as I don't think we'd want to pass mutable message state (the DOMWriteTx instances) - that would go against akka principals. We'd have to pass the IMDS instance to the helper actor on creation so it could create the write Tx's - but I'm not sure if that  would be a good idea. The helper actor would then have to return the mutable write Tx instances via a message or commit them when ask to do so.

Patch-set: 6