Update patch set 6
Patch Set 6:

I was thinking that worker actors only gets the batch to work on (no tx) and it would create an ImmutableCompositeModification from the batch and return back to Shard. Shard then creates a "WriteOnlyTx" and commits this ImmutableCompositeModification. So you dont have to share state between actors.

Thinking more about the parallelism part, I am not too sure if it will help. Bulk of the work is in reading from disk (that akka does) and commiting Tx (that happens sequentially). Deserialization is other piece that could consume time, but you have optimized it with your fixes. Here again, its sequential within the batch.

What do you think?

Patch-set: 6