Update patch set 6
Patch Set 6:

Initially I thought it was the Tx commit that was the bottleneck but it turned out to be deserialization. In my testing with Jan's flow script and batch size 5000, with the flow add batches, the commits  were only about 40-50 ms while deserialization took 4-5 s. After fixing the PreCondition inefficiencies and making other optimizations, deserialization is much improved but still takes the majority of the time. Applying the deserialized mods to the Tx takes about the same time as the commit from what I recall. I wanted to parallelize as much as I could.  The ordering of the commits has to be preserved and thus can't be done in parallel.

While using actors is an interesting idea, I don't see a compelling reason to take the time to re-implement it at this point. It works the way it is and we're running out of time for Helium and there's still other bugs to fix. Moiz said the serialization code isn't as performant as we would like so I want to look into that next.

BTW - for before and after comparisons in my testing running Jan's script to create and delete 15K flows, it takes about 12 s to recover the 30K journal entries, an improvement from 21 s prior. With the snapshot batch count at the default 20K, it takes a total of ~8 s, with the 1 snapshot and 10K journal entries. This is an improvement from 25 s.

Patch-set: 6