Update patch set 6
Patch Set 6:

It makes sense to parallelize if deser is the bottleneck. BTW, great improvement in performance.

My argument to use actors was to keep homogeneity, since we have been using actor as unit of concurrency. 

But I'd still go with +1 as it works and improves performance :-) Great job!

Patch-set: 6