Update patch set 2
Patch Set 2: Code-Review-1

(1 comment)

Concurrency issue on "close" method. Other than that looks good. I am running the test in a loop (back to back 100 times) to see if I can get the bug to reproduce, but at this point I have not been able to. 

The fix seems to line up with when I was seeing the problem. That is, the test failed when pending message count was full exceptions were printed, which makes perfect sense from a timing issue (that is if hte read thread is slower than the writer thread it would fail!).

The loop test is still running, but I think it should pass. So if we fix the small concurrency issue we can move forward and push this patch.

Patch-set: 2
Label: Code-Review=-1