Update patch set 2
Patch Set 2:

Moiz - I was able to get the shard recovery test working with a mock snapshot store and journal. See ShardTest#testRecovery in https://git.opendaylight.org/gerrit/#/c/11136/. I was able to derive from Shard with an associated Creator class and override onRecoveryComplete to update a CountDownLatch when complete and wait for the latch in the test then verify.

Patch-set: 2