Update patch set 2
Patch Set 2:

In looking at the log attachments in the bug, it seems this patch is to address the transaction commit failures due to "Could not find leader so transaction cannot be created" errors when trying to create the shard transaction. However it appears in looking at the changes, if the shard isn't initialized yet, we'll still get a transaction commit failure but with a different cause (i.e. NotInitializedEx).  The caller may or may not ever retry it.

I think the infrastructure you added such that callers can detect when a shard is present but not yet available is good but it seems to me the TransactionProxy could wait until the shard is available instead of failing fast with an exception. This would make it more resilient on startup especially since it could take significant time for a shard to recover from persistence. 

So I would propose to have the TransactionProxy treat NotInitializedEx as a transient error and continue retrying (after a sleep, say 50 ms) until it succeeds or some other unrecoverable error occurs. I think this is reasonable since we block the caller anyway when creating a shard transaction.

Patch-set: 2