Update patch set 2
Patch Set 2:

The FindLocalShard message can return ActorNotInitialized, in which case ActorContext#findLocalShard returns absent. In DistributedDataStore#registerChangeListener, if findLocalShard returns absent, the registration essentially fails by creating a NoOpDataChangeListenerRegistration (there's no recovery from this). I think this is problematic. 

As I mentioned in my previous comment, I think we need to handle ActorNotInitialized with retries. Here's my thoughts:

ActorContext#findLocalShard and ActorContext#findPrimaryShard should both throw NotInitializedException. 

In DistributedDataStore#registerChangeListener, always create a DataChangeListenerRegistrationProxy.  Move the code that performs the findLocalShard and RegisterChangeListener operations into DataChangeListenerRegistrationProxy. In DataChangeListenerRegistrationProxy, do the findLocalShard operation async. On success, then send the RegisterChangeListener message. On NotInitializedException, schedule a retry - we can use the ActorSystem's scheduler for this.

In TransactionProxy#createTransactionIfMissing we also need to retry (but synchronously).

There's one other startup timing case that was introduced by Moiz's recent work to delay creating the local shards in the ShardManager until it receives a complete SchemaContext. If there is a shard configured but it hasn't been created yet then findLocalShard (and maybe findPrimaryShard) would return LocalShardNotFound. I think the ShardManager should initialize the localShards map in the ctor but initially with no shard ActorRef in the ShardInformation (and shardInitialized = false). Then in updateSchemaContext, either create the shard actor if not yet set in the ShardInformation or send the the UpdateSchemaContext message.

Patch-set: 2