Update patch set 2
Patch Set 2: Code-Review-1

While reading from the in-memory store is cheap, reading from the distributed store carries overhead, especially if the data is on a remote shard. Instead of pre-reading, I would suggest to subclass a NodeDoesExistException from TransactionCommitFailedException and translate the internal ModifiedNodeDoesNotExistException to NodeDoesExistException. Then RestconfImpl can catch NodeDoesExistException and convert to a RestconfDocumentedException with RestconfError.DATA_MISSING.

Patch-set: 2
Label: Code-Review=-1