Update patch set 2
Patch Set 2:

Tom,

I will update ActorContext#findLocalShard to throw NotInitializedException.

Regarding retries, we could re-try for a fixed amount of time. After which we may succeed or still fail. If we fail, it will not be any different than how it is now. Also, it wont be wise to just block until datastore comes up as this may cause deadlocks.  
IMHO, retry logic is better left to client. Based on the need, client may continue to retry until it succeeds or try later on demand.
We could also think about evolving the APIs to provide notification on shard initialization to aid client in retries. OSGI level mechanics on this would be better. I think more thought and design needs to go into this.
This patch's main intent was to fail fast and provide a better exception message. We should consider evolving this in subsequent release. We wont have time to change now and provide satisfactory level of testing before RC2.

Patch-set: 2
Label: Verified=0