Update patch set 2
Patch Set 2:

Re: "it will not be any different than how it is now", the difference would be that with retries it will likely succeed and every attempt will have been made to satisfy the client's request.

With DCL registration, there is no error reporting in the API contract - clients don't expect it to fail. It's more or less a "set and forget" operation. As I outlined in my previous comment, we can make this completely async with retries without the client being aware.

For transactions, I think we should keep the concept of shards and their internal dynamics as transparent as possible to clients, i.e. make it easy as possible for clients to interact with the data store and let them focus on their business logic. Failing fast and expecting clients to handle retries does not make for easy data store interaction. We've already seen this with the OptimisticLockFailedException. It's hard enough for some people to deal with the async nature of the API - making them handle retries and having them worry about shard initialization would just make the API that much harder to use.

The only remaining part of the transaction process that is blocking is the initial  creation of the transaction. Blocking threads in this manner doesn't cause deadlocks - it's just not efficient for performance and parallelism. I've been meaning to make the transaction creation async, just haven't gotten to it.  If we do that then we can also handle shard lookup retries async w/o blocking.

I agree we should look into this post-RC2. I'll create bugs for tracking.

Patch-set: 2
Label: Code-Review=0