Update patch set 2
Patch Set 2:

About retries, if we dont guarantee success (which we cant), the client would still need to write code to handle failure, no matter how slim the chances are. In that respect, it wont be any different for the client.

I agree clients needn't be aware of concepts of shards and internal mechanics.  
This in my view is the case of OSGI module not ready for business, which is why I said we should handle "ready" notification OSGI way.

Blocking the thread would make client app not initialize. Topology manager, for eg, expects datastore during initialization. If there exists any such app with yang module and it doesnt get initialized, SchemaService will not be able to scan its yang file, while datastore might be waiting for SchemaContext to contain this yang module.....and so deadlock. To be honest, I need to check yangtools to see how this mechanics work exactly.

Patch-set: 2