Update patch set 3
Patch Set 3:

Abhishek - re: SchemaService, the ModuleInfoBundleTracker listens for ACTIVE bundles and scrapes the yang files to build up the SchemaContext. So you are correct that a form of deadlock could occur if a bundle has a bundle Activator and accesses the data store directly from the Activator (bad practice) and the bundle also supplies a yang model. The stars would have to align just right but it's theoretically possible. That kind of deadlock would occur no matter who is doing the retrying - the CDS or the client.

Re: retries: yes - clients need to handle Tx failures as the API has checked exceptions but most if not all clients will assume a hard, unrecoverable failure and just log the error and bail. They're not likely to retry. I don't think adding another constraint to the API for transient, recoverable error scenarios is going to help API consumption and usage. So IMO, keep it simple for clients ("dumb it down") and don't expose transient issues.

Patch-set: 3