Update patch set 1
Patch Set 1:

Ed and I discussed this - I agree with him that we should have authorization on by default. While it would be great if the controller didn't have to depend on AAA the reality is that it already does for other reasons.

By turning authorizaton on by default we provide by default what more than 80% of our customers would expect - that rest conf is secure and not open to the world. :)

+1 on trying to get this into RC2 respin.

Patch-set: 1