Update patch set 3
Patch Set 3:

Re: deadlock, my point was about not blocking the call. The right approach IMO is to use notifications and react.

Also, I am all for keeping it simple for clients but in a distributed environment, clients will need to be "more" aware of the system, as components may not be available all the time. Also, even with your proposed plan (to retry in CDS), you can not guarantee success and as a result client will still need to handle the failure. The system currently throws TimeoutException, even without my change. Clients will have to deal with it. I am certainly not opposed to retries. They will increase the likelihood of success but also add a degree of indeterminism in the system...sometimes fail, sometimes pass which will make it hard to diagnose problems. Needless to say, it will increase complexity in the code. 

Your point is valid about Tx API. I am throwing unchecked exception that clients may likely ignore. Do you think making it a checked exception, a better idea? About adding resiliency in such transient condition, lets discuss more on this with another defect? Without my changes (failing fast), the Shard was stashing all these Tx messages that would have been played onRecoverComplete. But ActorContext times it out which is what client saw. Do we just increase that timeout when actor is NOT initialized?

Patch-set: 3