BUG 720 - YANG leaf as JSON input *<*:* couldn't be saved
If input value for leaf in JSON format contained opending angle brackeg
(<) followed by collon (:) then exception was raised.

Because schema isn't present at time of JSON input reading it was also
supposed that data could be of type leafref which is specified in format
module:value. Module part is then transformed to instance of URI object.

Chracter < isn't permitted in URI according RFC2396 for URI and while
creating instance of URI class the exception which wasn't caught was
raised and transformed to returning error message.

Change-Id: I46bb949b38623f62a02daf4390c373371775ae1f
Signed-off-by: Jozef Gloncak <jgloncak@cisco.com>