Update patch set 1
Patch Set 1: Code-Review-1

Giovanni, I don't think this is the fix for this problem. I think the problem is that both clustering.stub and clustering.services-implementation were added as dependency to the adsal feature. Since both of these bundles are loaded and they implement the same interface some of the cache creations go to the stub. Subsequently when we try to add listeners the calls go to the services-implementation. This is why registrations are not working.

Patch-set: 1
Reviewer: Gerrit User 795 <795@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=-1