BUG 2074 : clustering.services-implementation failed to register listner.
The adsal feature depends on both the stub and implementaton of clustering.
This causes a problem when deployed because sometimes the stub is wired in first
which leads to createCache requests going to that implementation. Subsequently
the real implementation gets wired in and add listener calls go there but since
the caches were created on the stub this fails.

This patch removes the stub as a dependency with the assumption that it is not
required in production.

Change-Id: I42203bbf31b46560c23f283ddd83e56bd950830b
Signed-off-by: Moiz Raja <moraja@cisco.com>