Bug 2106: Refactored FlowCapableInventoryProvider to properly work with failures.
FlowCapableInventoryProvider was unable to continue if transaction chain
failed during construction of new transaction.

Refactored it in way, that it is able to reconstruct and retry transaction,
which failed because of previous transaction and retries it once
in new transaction chain. If it fails again, it continues with next batch.

Change-Id: I4a90a3b60b0a49d562c8648a779e790f55840b5d
Signed-off-by: Tony Tkacik <ttkacik@cisco.com>