BUG 906 - pagination and searching in swagger
It was reported that pagination and searching should be available to
decrease amount of data which have to be transfered when REST API for all
YANG modules is generated.

After GET on
http://server:port/apidoc/apis/modules
the list of yang modules will be returned. Every listed module contains at
least one rpc or list|container as top level. It is also possible to
search via searching area in top right corner. It is also possible to
display pagination (currently it is switched off).

Every module name in table is hyperlink
http://server:port/apidoc/explorer?moduleName=name&revision=YYYY-DD-MM
It will navigate swagger UI where only API links for this module are loaded.

Original access to REST apis kept preserved. It means that is is still
possible to list all REST apis via
http://server:port/appidoc/explorer

Change-Id: Ic848c71f8ba1b24ac835865a283530b6daac2859
Signed-off-by: Jozef Gloncak <jgloncak@cisco.com>