[DO NOT MERGE] Incrementing versions by 0.0.1 for post-Helium stability branch
This patch should be a draft only to see if the config persister test failure
is a heisenbug or not.

Excluding the aaa version as there is a cyclic dependency there.
A second patch will update this version. Also, this test temporarily
stops running the netconf-connector and restconf feature tests. This
will also be re-enabled in the same second patch.

Change-Id: Id33499fe343090a2c38b22038eefdd1bb13f4291
Signed-off-by: Colin Dixon <colin@colindixon.com>