Update patch set 2
Patch Set 2:

I am fairly certain that this is an honest failure where controller is incompatible with 0.7.0 yangtools, but was with 0.6.2 yangtools.

If Robert or Tony get a chance to look at it and fix things, that would be great. Otherwise, I might or might not get to this until Monday.

Patch-set: 2
Label: Verified=0