Bug 2135: Create ShardInformation on startup
Modified the ShardManager to create the ShardInformation map on startup,
but without creating the shard actors yet, instead of when the SchemaContext is initialized.

When the SchemaContext is fully initialized then create the shard actors.

On FindLocalShard and FindPrimaryShard messages, only return a valid
response if the shard actor is created and it's initialized.

Change-Id: I361f2af9e53878f62c5890350afd7a5f2877b95c
Signed-off-by: tpantelis <tpanteli@brocade.com>