Removed cyclic dependency on AAA
Cyclic dependency was only introduced in Karaf features,
where actual authentification code for Netconf and Restconf
was present in AAA project, so it is natural that this
features which also include authentification should
live in AAA and/or integration/packaging project.

This was suggested before cycle was introduced in code
late in release cycle in Bug 1948.

Change-Id: I0db9b6012bfb50fb58bb81d777f4c346f0f5ba04
Signed-off-by: Tony Tkacik <ttkacik@cisco.com>