BUG-2838: Trim leading whitespaces from incoming NETCONF messages.
Some network devices (such as Cisco routers) send RPC replies with a leading
newline. The OpenDaylight controller does not accept those replies: the XML
parsing fails.

This patch fixes the NETCONF messages before they are sent to the XML
parser: it removes all the whitespaces at the beginning of the messages.

Change-Id: Ibc6eb6dc5bad6252a3c9bed73d3db83814aff501
Signed-off-by: Gwenael Lambrouin <gwenael.lambrouin@b-com.com>