Update patch set 1
Patch Set 1:

Can you clarify what the outcome of this patch is? Is it just that we won't get an exception now when a yang like opflex.yang is in the list?

Asked another way, how should someone test and verify this?

I copied opflex.yang to toaster, rebuilt and deployed and while I don't see opflex.yang in the restconf doc, but restconf docs do render without error.

Patch-set: 1