Update patch set 2
Patch Set 2: Code-Review-1

The patch seems fine, but can you add / enhance a unit test to cover this scenario? I don't want us to have this functionality removed. A unit test would define the expect contract to make sure we don't loose this functionality down the road. If it is already covered please point me to the test.

Also, I would add a comment in the code about the edge case we are catching here. Otherwise people coming later might wonder why we essentially catch and hide what may seem like a valid exception.

I would prefer these changes are made with this patch. If it is critical that this patch gets through we can submit it, but I would like to see a new bug opened to address the above enhancements and followed up on shortly after.

Thank you

Patch-set: 2
Label: Code-Review=-1