Update patch set 1
Patch Set 1:

Mark, I believe the original intent of the bug was to eliminate the need for RaftState by correctly implementing the State pattern. So the basic idea is that RaftBehavior#handleMessage should have returned RaftBehavior instead of RaftState. 

At the end of the day I don't see a need to remove RaftState altogether and I see you have not done so.

Patch-set: 1
Label: Code-Review=0