Update patch set 3
Patch Set 3:

Moiz - I started going down the path of creating our own broker but realized it was going to be a bit of a pain in addition to duplication of code. We would need our own yang module and related code etc. By re-using the existing broker we get all the MBean stats for free and we'll get whatever future extensions are made to the existing broker.  

My take on the intent of the bug was to gain shard commit concurrency between transactions so that was what I focused on. To do that we just need slightly different coordination behavior (basically not single-thread the commits) so it seemed reasonable and simpler to me to just configure a different coordination impl in the existing data broker.    

Coordination between config/operational datastore would only come into play if a single transaction modified both stores but that would be very uncommon. However I think coordinating between the 2 is still technically correct based on the API contract. 

We can certainly revisit this if you want.

Patch-set: 3
Reviewer: Gerrit User 1842 <1842@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0