Update patch set 2
Patch Set 2:

It seems that with configuring the ActorSystem with a non-persistent journal we lose persistence for the ShardManager as well (the schema context modules).  How about if we always use the persistent journal, as before, and add the persistent flag to the RaftActor. So in appendAndPersist, if persistent is false, skip the super.persist call but still execute the code in the Procedure#apply method. Also during recovery, if persistent is false, ignore the messages (and clear the journal/snapshot when recovery complete).

Patch-set: 2