Update patch set 2
Patch Set 2:

The only reason we needed to persist the SchemaContext data in ShardManager was because we wanted to gate access to the Shard if the SchemaContext was not right. When the data itself is not persistent we do not need to gate creation of the Shard because the Shard will always come up empty.

Patch-set: 2