Update patch set 2
Patch Set 2:

Yes - you're right - we only needed persistence in the ShardManager to avoid Shard persistence recovery errors.

Still, I think disabling persistence for the entire data store's ActorSystem might be fragile going forward. If we ever need persistence for any other actor in the future that is unrelated to the Shard, we would have a problem. Also I'm still not clear on why we need the NonPersistentJournal to keep the last entry index. Is that specific to how the Shard needs it to behave? Just wondering if NonPersistentJournal is general or tailored to the Shard...

Also I think eventually we will need an ActorSystem per data store to support reconfiguring/restarting each data store independently  via the config system. For example, one may want to change a config param for the operational data store on the fly w/o restarting the controller. In that case just the operational store module would be stopped and new instance spun up. On close, the DistributedDatastore shuts down its ActorContext instance which shuts down the ActorSystem. However, if the ActorSystem is shared with the config data store then that would hose the config store. 

So, since this patch is adding the ability for multiple ActorSystems, can we just create one per data store instead of re-using persistent/non-persistent ActorSystems across data stores?

Patch-set: 2
Label: Code-Review=0