Update patch set 2
Patch Set 2:

(5 comments)

The purpose of swapping journal/snapshot at the actor system is to least impact our Raft code which should really not be even concerned with the datastore. It keeps the design clean. 

I agree with you that it's a little fragile in that you cannot have some actors persistent and some not.  Apparently there are hooks in akka-persistence to allow multiple journals/snapshot stores per actor system https://github.com/akka/akka/issues/15587. If that became available we may not even need the additional actor system.

As for separate actor system for config/operational datastore. I like the idea. In fact if you agree we could simply ditch enabling/disabling persistent using the persistent flag in config-subsystem and simply do it based on the journal and snapshot configuration in akka config.

In Lithium we will need to refactor Shard/RaftActor to allow for other "clustering strategies" we can do the work on enabling/disabling persistence per shard then.

Patch-set: 2
Label: Verified=0