Update patch set 3
Patch Set 3:

(1 comment)

A restart would be required. I do not think it makes sense for these configurations to be modified in a running production system so I am not really that concerned about this fact. 

The only reason I bring up this is because with the persistent flag + different actor system for each data store we would need 4 separate akka configurations.

Patch-set: 3