Update patch set 3
Patch Set 3:

Ok - if using akka config makes it easier than I'm fine with it.

One final point of discussion... if we think we may have uses cases in the future to support enabling/disabling persistence per shard, would we be able to accomplish it with this approach (i.e. configuring it at the actor system level)? I want to make sure we're not painting ourselves into a corner. I'm still wondering if handling it in the RaftActor via a persistent flag, while maybe not the most elegant solution, may give us the most flexibility.

Patch-set: 3