Update patch set 3
Patch Set 3:

Moiz - I'm kind of torn on it. There's advantages/disadvantages either way. If we do the conditional logic in the RaftActor, it sets us up nicely for per-shard persistence in the future. It also avoids the potential issue of interfering with another actor's persistence (even though that's not an issue now). 

On the other hand, doing it at the journal store level makes the RaftActor cleaner although it may not help us if/when we need per-shard persistence. Also akka would still be able to get the highest sequence number with your NonPersistentJournal impl although I'm not sure what we would lose without that.

I could go either way but I'm leaning towards the conditional approach (if we don't really need the highest sequence number). I'd be interested in other's opinions....

Either way I think we should create an ActorSystem per data store but if we go with the conditional approach then that probably should be done with another bug/patch.

Patch-set: 3
Label: Verified=0